Separate indexing and searching of documentation
