+++ app/Commands/Docs/IndexAllDocuments.php
@@ -1,7 +1,7 @@
 <?php  namespace App\Commands\Docs;
 
 use App\Commands\Command;
-use App\DocSearchService;
+use App\Services\Documentation\Indexer;
 use Illuminate\Contracts\Bus\SelfHandling;
 
 class IndexAllDocuments extends Command implements SelfHandling {
@@ -16,10 +16,10 @@ class IndexAllDocuments extends Command implements SelfHandling {
     /**
      * Execute the command.
      *
-     * @return void
+     * @param Indexer $indexer
      */
-    public function handle(DocSearchService $searcher)
+    public function handle(Indexer $indexer)
     {
-        $searcher->indexAllDocuments();
+        $indexer->indexAllDocuments();
     }
 }
