+++ app/Commands/Docs/InitializeElasticSearchIndex.php
@@ -1,7 +1,7 @@
 <?php  namespace App\Commands\Docs;
 
 use App\Commands\Command;
-use App\DocSearchService;
+use App\Services\Documentation\Indexer;
 use Illuminate\Contracts\Bus\SelfHandling;
 
 class InitializeElasticSearchIndex extends Command implements SelfHandling {
@@ -16,10 +16,10 @@ class InitializeElasticSearchIndex extends Command implements SelfHandling {
     /**
      * Execute the command.
      *
-     * @return void
+     * @param Indexer $indexer
      */
-    public function handle(DocSearchService $searcher)
+    public function handle(Indexer $indexer)
     {
-        $searcher->initializeIndex();
+        $indexer->initializeIndex();
     }
 }
