+++ app/Services/Documentation/Searcher.php
@@ -1,0 +1,54 @@
+<?php namespace App\Services\Documentation;
+
+use Elasticsearch\Client;
+use Elasticsearch\Common\Exceptions\Missing404Exception;
+
+class Searcher {
+
+	/**
+	 * @var Client
+	 */
+	private $client;
+
+	public function __construct(Client $client)
+    {
+		$this->client = $client;
+	}
+
+	/**
+	 * @todo Make the search broader--for example, include and value more highly the title
+	 * @todo Make the search smarter--for example, care more about h2s, h3s, etc.
+	 * @param  string $version
+	 * @param  string $term
+	 * @return array
+	 * @throws \Exception
+	 */
+	public function searchForTerm($version, $term)
+	{
+		$params['index'] = $this->getIndexName($version);
+		$params['type'] = 'page';
+		$params['body']['query']['match']['body.md'] = $term;
+
+		try {
+			$response = $this->client->search($params);
+		} catch (Missing404Exception $e) {
+			throw new \Exception('ElasticSearch Index was not initialized.');
+		}
+
+		// @todo Validate response
+
+		return $response['hits']['hits'];
+	}
+
+	/**
+	 * Get the ElasticSearch index name for this version
+	 *
+	 * @todo Duplicated to Indexer; fix
+	 * @param $version
+	 * @return string
+	 */
+	private function getIndexName($version)
+	{
+		return 'docs.' . $version;
+	}
+}
